+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Forma 1.0.0
+
+* Added `subParser` and `field''` combinators. Changed signature of
+  `mkFieldError` to accept `NonEmpty (SelectedName names)` as first
+  argument.
+
 ## Forma 0.2.0
 
 * Added `withCheck`.